ContentHandler: Moved amazon parser and content_handler to subfolder
